Set localtime to Chicago.
This is configured as a symlink, so this is really just a
fake configuration that helps me keep track of changes that
I do actually make to localtime.

I expect I'll have to do this sort of thing with other
configurations too, like systemd services and Sublime Text
installed packages, which I can't easily version control.

I actually live in Champaign right now, but Chicago is the
closest, and I'm in Chicago a lot anyway.