Add a home directory for all my code repositories.
That's where this repository is! :D

I used to use ~/dev, but that was confusing because /dev.

I don't think that I could get clever enough with
git submodules to actually link to my repos here. It
will probably stay empty in this repository even though
my code is actually here.