Install i3-wm, i3status, and dmenu.
dmenu is configured as the program starter by default on i3. I
don't know if there's anything better. It will certainly work
for now.

i3lock is also in the i3 group, but I don't like it.

The config is the default generated by the first-time pop-up. I
selected my Windows key as the super key.