Install chromium.
I need to extend this with the internal pdf viewer from the AUR
and with a flash player.

I'm surprised at how many dependencies this pulled in. It even
included Python 3.

I've heard of more minimal browsers, but I don't think I'm ready
to switch yet. A minimal theme for Chromium that blends in with
i3 would be nice though.