Add OpenVPN.
This definitely needs to be configured extensively so that
VPN connections are as easy to bring up and down as wifi
networks.

There might be DNS leaks, which need to be plugged ASAP. Maybe
you could also do some sort of filtering to ensure there are no
leaks.

It would also be useful to show my VPN's status in the status bar.